+++ pkg/factory/scheduler/enablement_test.go
@@ -4,87 +4,11 @@ import (
 	"context"
 	"strings"
 	"testing"
-	"time"
 
 	"github.com/portpowered/infinite-you/pkg/factory/state"
 	"github.com/portpowered/infinite-you/pkg/interfaces"
 	"github.com/portpowered/infinite-you/pkg/petri"
 )
-
-// capturingLogger records all log calls for assertion.
-type capturingLogger struct {
-	entries []logEntry
-}
-
-type logEntry struct {
-	level string
-	msg   string
-	args  []any
-}
-
-func (l *capturingLogger) Debug(msg string, keysAndValues ...any) {
-	l.entries = append(l.entries, logEntry{level: "debug", msg: msg, args: keysAndValues})
-}
-func (l *capturingLogger) Info(msg string, keysAndValues ...any) {
-	l.entries = append(l.entries, logEntry{level: "info", msg: msg, args: keysAndValues})
-}
-func (l *capturingLogger) Warn(msg string, keysAndValues ...any) {
-	l.entries = append(l.entries, logEntry{level: "warn", msg: msg, args: keysAndValues})
-}
-func (l *capturingLogger) Error(msg string, keysAndValues ...any) {
-	l.entries = append(l.entries, logEntry{level: "error", msg: msg, args: keysAndValues})
-}
-func (l *capturingLogger) Verbose(msg string, keysAndValues ...any) {
-	l.entries = append(l.entries, logEntry{level: "verbose", msg: msg, args: keysAndValues})
-}
-
-func (l *capturingLogger) entryMatches(e *logEntry, substr string) bool {
-	if strings.Contains(e.msg, substr) {
-		return true
-	}
-	for _, arg := range e.args {
-		if s, ok := arg.(string); ok && strings.Contains(s, substr) {
-			return true
-		}
-	}
-	return false
-}
-
-func (l *capturingLogger) findEntry(substr string) *logEntry {
-	for i := range l.entries {
-		if l.entryMatches(&l.entries[i], substr) {
-			return &l.entries[i]
-		}
-	}
-	return nil
-}
-
-func (l *capturingLogger) countEntries(substr string) int {
-	count := 0
-	for i := range l.entries {
-		if l.entryMatches(&l.entries[i], substr) {
-			count++
-		}
-	}
-	return count
-}
-
-func makeTestSnapshot(tokens map[string]*interfaces.Token) petri.MarkingSnapshot {
-	placeTokens := make(map[string][]string)
-	for id, tok := range tokens {
-		if tok.CreatedAt.IsZero() {
-			tok.CreatedAt = time.Now()
-		}
-		if tok.EnteredAt.IsZero() {
-			tok.EnteredAt = time.Now()
-		}
-		placeTokens[tok.PlaceID] = append(placeTokens[tok.PlaceID], id)
-	}
-	return petri.MarkingSnapshot{
-		Tokens:      tokens,
-		PlaceTokens: placeTokens,
-	}
-}
 
 func TestEnablementEvaluator_LogsEnabledTransition(t *testing.T) {
 	logger := &capturingLogger{}
@@ -115,15 +39,10 @@ func TestEnablementEvaluator_LogsEnabledTransition(t *testing.T) {
 		t.Fatalf("expected 1 enabled transition, got %d", len(enabled))
 	}
 
-	// Should have logged the enabled transition.
-	entry := logger.findEntry("transition enabled")
-	if entry == nil {
+	if entry := logger.findEntry("transition enabled"); entry == nil {
 		t.Fatal("expected 'transition enabled' log entry")
 	}
-
-	// Should have the summary entry.
-	summary := logger.findEntry("evaluation complete")
-	if summary == nil {
+	if summary := logger.findEntry("evaluation complete"); summary == nil {
 		t.Fatal("expected 'evaluation complete' log entry")
 	}
 }
@@ -148,14 +67,10 @@ func TestEnablementEvaluator_LogsDisabledInsufficientTokens(t *testing.T) {
 	}
 
 	marking := makeTestSnapshot(map[string]*interfaces.Token{})
-
-	enabled := eval.FindEnabledTransitions(context.Background(), n, &marking)
-	if len(enabled) != 0 {
+	if enabled := eval.FindEnabledTransitions(context.Background(), n, &marking); len(enabled) != 0 {
 		t.Fatalf("expected 0 enabled transitions, got %d", len(enabled))
 	}
-
-	entry := logger.findEntry("transition disabled")
-	if entry == nil {
+	if entry := logger.findEntry("transition disabled"); entry == nil {
 		t.Fatal("expected 'transition disabled' log entry")
 	}
 }
@@ -194,14 +109,10 @@ func TestEnablementEvaluator_LogsDisabledGuardFailed(t *testing.T) {
 		"tok-review": {ID: "tok-review", PlaceID: "p-review", Color: interfaces.TokenColor{WorkID: "r1", ParentID: "WRONG"}},
 	})
 
-	enabled := eval.FindEnabledTransitions(context.Background(), n, &marking)
-	if len(enabled) != 0 {
+	if enabled := eval.FindEnabledTransitions(context.Background(), n, &marking); len(enabled) != 0 {
 		t.Fatalf("expected 0 enabled transitions, got %d", len(enabled))
 	}
-
-	// Should log guard failure reason.
-	entry := logger.findEntry("guard failed")
-	if entry == nil {
+	if entry := logger.findEntry("guard failed"); entry == nil {
 		t.Fatal("expected log entry containing 'guard failed'")
 	}
 }
@@ -306,42 +217,17 @@ func TestEnablementEvaluator_BindsAllTokensForMatchingParentGuard(t *testing.T) 
 	if len(enabled) != 1 {
 		t.Fatalf("enabled transitions = %d, want 1", len(enabled))
 	}
-	gotParent := tokenIDs(enabled[0].Bindings["parent"])
-	if strings.Join(gotParent, ",") != "tok-parent" {
-		t.Fatalf("parent binding tokens = %v, want [tok-parent]", gotParent)
+	if got := tokenIDs(enabled[0].Bindings["parent"]); strings.Join(got, ",") != "tok-parent" {
+		t.Fatalf("parent binding tokens = %v, want [tok-parent]", got)
 	}
-	gotChildren := tokenIDs(enabled[0].Bindings["children"])
-	if strings.Join(gotChildren, ",") != "tok-child-a,tok-child-b" {
-		t.Fatalf("children binding tokens = %v, want [tok-child-a tok-child-b]", gotChildren)
+	if got := tokenIDs(enabled[0].Bindings["children"]); strings.Join(got, ",") != "tok-child-a,tok-child-b" {
+		t.Fatalf("children binding tokens = %v, want [tok-child-a tok-child-b]", got)
 	}
 }
 
 func TestEnablementEvaluator_SameNameGuardEnablesOnMatchingNames(t *testing.T) {
 	eval := NewEnablementEvaluator(nil)
-
-	n := &state.Net{
-		Places: map[string]*petri.Place{
-			"plan:ready": {ID: "plan:ready"},
-			"task:ready": {ID: "task:ready"},
-		},
-		Transitions: map[string]*petri.Transition{
-			"match-items": {
-				ID:   "match-items",
-				Name: "match-items",
-				InputArcs: []petri.Arc{
-					{ID: "plan-in", Name: "plan", PlaceID: "plan:ready", Direction: petri.ArcInput, Cardinality: petri.ArcCardinality{Mode: petri.CardinalityOne}},
-					{
-						ID:          "task-in",
-						Name:        "task",
-						PlaceID:     "task:ready",
-						Direction:   petri.ArcInput,
-						Cardinality: petri.ArcCardinality{Mode: petri.CardinalityOne},
-						Guard:       &petri.SameNameGuard{MatchBinding: "plan"},
-					},
-				},
-			},
-		},
-	}
+	n := sameNameGuardNet()
 	marking := makeTestSnapshot(map[string]*interfaces.Token{
 		"plan-alpha": {ID: "plan-alpha", PlaceID: "plan:ready", Color: interfaces.TokenColor{Name: "alpha"}},
 		"task-alpha": {ID: "task-alpha", PlaceID: "task:ready", Color: interfaces.TokenColor{Name: "alpha"}},
@@ -362,30 +248,7 @@ func TestEnablementEvaluator_SameNameGuardEnablesOnMatchingNames(t *testing.T) {
 
 func TestEnablementEvaluator_SameNameGuardBlocksNonMatchingNames(t *testing.T) {
 	eval := NewEnablementEvaluator(nil)
-
-	n := &state.Net{
-		Places: map[string]*petri.Place{
-			"plan:ready": {ID: "plan:ready"},
-			"task:ready": {ID: "task:ready"},
-		},
-		Transitions: map[string]*petri.Transition{
-			"match-items": {
-				ID:   "match-items",
-				Name: "match-items",
-				InputArcs: []petri.Arc{
-					{ID: "plan-in", Name: "plan", PlaceID: "plan:ready", Direction: petri.ArcInput, Cardinality: petri.ArcCardinality{Mode: petri.CardinalityOne}},
-					{
-						ID:          "task-in",
-						Name:        "task",
-						PlaceID:     "task:ready",
-						Direction:   petri.ArcInput,
-						Cardinality: petri.ArcCardinality{Mode: petri.CardinalityOne},
-						Guard:       &petri.SameNameGuard{MatchBinding: "plan"},
-					},
-				},
-			},
-		},
-	}
+	n := sameNameGuardNet()
 	marking := makeTestSnapshot(map[string]*interfaces.Token{
 		"plan-alpha": {ID: "plan-alpha", PlaceID: "plan:ready", Color: interfaces.TokenColor{Name: "alpha"}},
 		"task-beta":  {ID: "task-beta", PlaceID: "task:ready", Color: interfaces.TokenColor{Name: "beta"}},
@@ -423,21 +286,9 @@ func TestEnablementEvaluator_SameNameGuardFindsLaterMatchingBinding(t *testing.T
 		},
 	}
 	marking := makeTestSnapshot(map[string]*interfaces.Token{
-		"task-alpha": {
-			ID:      "task-alpha",
-			PlaceID: "task:to-complete",
-			Color:   interfaces.TokenColor{Name: "alpha"},
-		},
-		"task-zeta": {
-			ID:      "task-zeta",
-			PlaceID: "task:to-complete",
-			Color:   interfaces.TokenColor{Name: "zeta"},
-		},
-		"idea-zeta": {
-			ID:      "idea-zeta",
-			PlaceID: "idea:to-complete",
-			Color:   interfaces.TokenColor{Name: "zeta"},
-		},
+		"task-alpha": {ID: "task-alpha", PlaceID: "task:to-complete", Color: interfaces.TokenColor{Name: "alpha"}},
+		"task-zeta":  {ID: "task-zeta", PlaceID: "task:to-complete", Color: interfaces.TokenColor{Name: "zeta"}},
+		"idea-zeta":  {ID: "idea-zeta", PlaceID: "idea:to-complete", Color: interfaces.TokenColor{Name: "zeta"}},
 	})
 
 	enabled := eval.FindEnabledTransitions(context.Background(), n, &marking)
@@ -449,294 +300,6 @@ func TestEnablementEvaluator_SameNameGuardFindsLaterMatchingBinding(t *testing.T
 	}
 	if got := tokenIDs(enabled[0].Bindings["idea"]); strings.Join(got, ",") != "idea-zeta" {
 		t.Fatalf("idea binding tokens = %v, want [idea-zeta]", got)
-	}
-}
-
-func TestEnablementEvaluator_MatchesFieldsGuardEnablesSingleInputWhenSelectorResolves(t *testing.T) {
-	eval := NewEnablementEvaluator(nil)
-
-	n := &state.Net{
-		Places: map[string]*petri.Place{
-			"task:ready": {ID: "task:ready"},
-		},
-		Transitions: map[string]*petri.Transition{
-			"match-single": {
-				ID:   "match-single",
-				Name: "match-single",
-				InputArcs: []petri.Arc{{
-					ID:          "task-in",
-					Name:        "task",
-					PlaceID:     "task:ready",
-					Direction:   petri.ArcInput,
-					Cardinality: petri.ArcCardinality{Mode: petri.CardinalityOne},
-					Guard:       &petri.MatchesFieldsGuard{InputKey: `.Tags["_last_output"]`},
-				}},
-			},
-		},
-	}
-	marking := makeTestSnapshot(map[string]*interfaces.Token{
-		"task-alpha": {
-			ID:      "task-alpha",
-			PlaceID: "task:ready",
-			Color: interfaces.TokenColor{Tags: map[string]string{
-				"_last_output": "alpha",
-			}},
-		},
-	})
-
-	enabled := eval.FindEnabledTransitions(context.Background(), n, &marking)
-	if len(enabled) != 1 {
-		t.Fatalf("enabled transitions = %d, want 1", len(enabled))
-	}
-	if got := tokenIDs(enabled[0].Bindings["task"]); strings.Join(got, ",") != "task-alpha" {
-		t.Fatalf("task binding tokens = %v, want [task-alpha]", got)
-	}
-}
-
-func TestEnablementEvaluator_MatchesFieldsGuardEnablesOnMatchingTwoInputValues(t *testing.T) {
-	eval := NewEnablementEvaluator(nil)
-
-	n := &state.Net{
-		Places: map[string]*petri.Place{
-			"plan:ready": {ID: "plan:ready"},
-			"task:ready": {ID: "task:ready"},
-		},
-		Transitions: map[string]*petri.Transition{
-			"match-pair": {
-				ID:   "match-pair",
-				Name: "match-pair",
-				InputArcs: []petri.Arc{
-					{
-						ID:          "plan-in",
-						Name:        "plan",
-						PlaceID:     "plan:ready",
-						Direction:   petri.ArcInput,
-						Cardinality: petri.ArcCardinality{Mode: petri.CardinalityOne},
-						Guard:       &petri.MatchesFieldsGuard{InputKey: `.Tags["_last_output"]`},
-					},
-					{
-						ID:          "task-in",
-						Name:        "task",
-						PlaceID:     "task:ready",
-						Direction:   petri.ArcInput,
-						Cardinality: petri.ArcCardinality{Mode: petri.CardinalityOne},
-						Guard: &petri.MatchesFieldsGuard{
-							InputKey:     `.Tags["_last_output"]`,
-							MatchBinding: "plan",
-						},
-					},
-				},
-			},
-		},
-	}
-	marking := makeTestSnapshot(map[string]*interfaces.Token{
-		"plan-alpha": {
-			ID:      "plan-alpha",
-			PlaceID: "plan:ready",
-			Color: interfaces.TokenColor{Tags: map[string]string{
-				"_last_output": "alpha",
-			}},
-		},
-		"task-alpha": {
-			ID:      "task-alpha",
-			PlaceID: "task:ready",
-			Color: interfaces.TokenColor{Tags: map[string]string{
-				"_last_output": "alpha",
-			}},
-		},
-		"task-beta": {
-			ID:      "task-beta",
-			PlaceID: "task:ready",
-			Color: interfaces.TokenColor{Tags: map[string]string{
-				"_last_output": "beta",
-			}},
-		},
-	})
-
-	enabled := eval.FindEnabledTransitions(context.Background(), n, &marking)
-	if len(enabled) != 1 {
-		t.Fatalf("enabled transitions = %d, want 1", len(enabled))
-	}
-	if got := tokenIDs(enabled[0].Bindings["plan"]); strings.Join(got, ",") != "plan-alpha" {
-		t.Fatalf("plan binding tokens = %v, want [plan-alpha]", got)
-	}
-	if got := tokenIDs(enabled[0].Bindings["task"]); strings.Join(got, ",") != "task-alpha" {
-		t.Fatalf("task binding tokens = %v, want [task-alpha]", got)
-	}
-}
-
-func TestEnablementEvaluator_MatchesFieldsGuardBlocksMismatchedTwoInputValues(t *testing.T) {
-	eval := NewEnablementEvaluator(nil)
-
-	n := &state.Net{
-		Places: map[string]*petri.Place{
-			"plan:ready": {ID: "plan:ready"},
-			"task:ready": {ID: "task:ready"},
-		},
-		Transitions: map[string]*petri.Transition{
-			"match-pair": {
-				ID:   "match-pair",
-				Name: "match-pair",
-				InputArcs: []petri.Arc{
-					{
-						ID:          "plan-in",
-						Name:        "plan",
-						PlaceID:     "plan:ready",
-						Direction:   petri.ArcInput,
-						Cardinality: petri.ArcCardinality{Mode: petri.CardinalityOne},
-						Guard:       &petri.MatchesFieldsGuard{InputKey: `.Tags["_last_output"]`},
-					},
-					{
-						ID:          "task-in",
-						Name:        "task",
-						PlaceID:     "task:ready",
-						Direction:   petri.ArcInput,
-						Cardinality: petri.ArcCardinality{Mode: petri.CardinalityOne},
-						Guard: &petri.MatchesFieldsGuard{
-							InputKey:     `.Tags["_last_output"]`,
-							MatchBinding: "plan",
-						},
-					},
-				},
-			},
-		},
-	}
-	marking := makeTestSnapshot(map[string]*interfaces.Token{
-		"plan-alpha": {
-			ID:      "plan-alpha",
-			PlaceID: "plan:ready",
-			Color: interfaces.TokenColor{Tags: map[string]string{
-				"_last_output": "alpha",
-			}},
-		},
-		"task-beta": {
-			ID:      "task-beta",
-			PlaceID: "task:ready",
-			Color: interfaces.TokenColor{Tags: map[string]string{
-				"_last_output": "beta",
-			}},
-		},
-	})
-
-	if enabled := eval.FindEnabledTransitions(context.Background(), n, &marking); len(enabled) != 0 {
-		t.Fatalf("enabled transitions = %d, want 0", len(enabled))
-	}
-}
-
-func TestEnablementEvaluator_MatchesFieldsGuardRequiresAllInputsToMatchSourceValue(t *testing.T) {
-	eval := NewEnablementEvaluator(nil)
-
-	n := &state.Net{
-		Places: map[string]*petri.Place{
-			"plan:ready":  {ID: "plan:ready"},
-			"task:ready":  {ID: "task:ready"},
-			"asset:ready": {ID: "asset:ready"},
-		},
-		Transitions: map[string]*petri.Transition{
-			"match-triplet": {
-				ID:   "match-triplet",
-				Name: "match-triplet",
-				InputArcs: []petri.Arc{
-					{
-						ID:          "plan-in",
-						Name:        "plan",
-						PlaceID:     "plan:ready",
-						Direction:   petri.ArcInput,
-						Cardinality: petri.ArcCardinality{Mode: petri.CardinalityOne},
-						Guard:       &petri.MatchesFieldsGuard{InputKey: `.Tags["_last_output"]`},
-					},
-					{
-						ID:          "task-in",
-						Name:        "task",
-						PlaceID:     "task:ready",
-						Direction:   petri.ArcInput,
-						Cardinality: petri.ArcCardinality{Mode: petri.CardinalityOne},
-						Guard: &petri.MatchesFieldsGuard{
-							InputKey:     `.Tags["_last_output"]`,
-							MatchBinding: "plan",
-						},
-					},
-					{
-						ID:          "asset-in",
-						Name:        "asset",
-						PlaceID:     "asset:ready",
-						Direction:   petri.ArcInput,
-						Cardinality: petri.ArcCardinality{Mode: petri.CardinalityOne},
-						Guard: &petri.MatchesFieldsGuard{
-							InputKey:     `.Tags["_last_output"]`,
-							MatchBinding: "plan",
-						},
-					},
-				},
-			},
-		},
-	}
-
-	matching := makeTestSnapshot(map[string]*interfaces.Token{
-		"plan-alpha": {
-			ID:      "plan-alpha",
-			PlaceID: "plan:ready",
-			Color: interfaces.TokenColor{Tags: map[string]string{
-				"_last_output": "alpha",
-			}},
-		},
-		"task-alpha": {
-			ID:      "task-alpha",
-			PlaceID: "task:ready",
-			Color: interfaces.TokenColor{Tags: map[string]string{
-				"_last_output": "alpha",
-			}},
-		},
-		"asset-alpha": {
-			ID:      "asset-alpha",
-			PlaceID: "asset:ready",
-			Color: interfaces.TokenColor{Tags: map[string]string{
-				"_last_output": "alpha",
-			}},
-		},
-		"asset-beta": {
-			ID:      "asset-beta",
-			PlaceID: "asset:ready",
-			Color: interfaces.TokenColor{Tags: map[string]string{
-				"_last_output": "beta",
-			}},
-		},
-	})
-
-	enabled := eval.FindEnabledTransitions(context.Background(), n, &matching)
-	if len(enabled) != 1 {
-		t.Fatalf("enabled transitions = %d, want 1", len(enabled))
-	}
-	if got := tokenIDs(enabled[0].Bindings["asset"]); strings.Join(got, ",") != "asset-alpha" {
-		t.Fatalf("asset binding tokens = %v, want [asset-alpha]", got)
-	}
-
-	mismatched := makeTestSnapshot(map[string]*interfaces.Token{
-		"plan-alpha": {
-			ID:      "plan-alpha",
-			PlaceID: "plan:ready",
-			Color: interfaces.TokenColor{Tags: map[string]string{
-				"_last_output": "alpha",
-			}},
-		},
-		"task-alpha": {
-			ID:      "task-alpha",
-			PlaceID: "task:ready",
-			Color: interfaces.TokenColor{Tags: map[string]string{
-				"_last_output": "alpha",
-			}},
-		},
-		"asset-beta": {
-			ID:      "asset-beta",
-			PlaceID: "asset:ready",
-			Color: interfaces.TokenColor{Tags: map[string]string{
-				"_last_output": "beta",
-			}},
-		},
-	})
-
-	if enabled := eval.FindEnabledTransitions(context.Background(), n, &mismatched); len(enabled) != 0 {
-		t.Fatalf("enabled transitions with mismatched third input = %d, want 0", len(enabled))
 	}
 }
 
@@ -774,9 +337,7 @@ func TestEnablementEvaluator_VisitCountGuardEnablesAtThreshold(t *testing.T) {
 		"tok-work": {
 			ID:      "tok-work",
 			PlaceID: "p-init",
-			History: interfaces.TokenHistory{
-				TotalVisits: map[string]int{"review": 2},
-			},
+			History: interfaces.TokenHistory{TotalVisits: map[string]int{"review": 2}},
 		},
 	})
 	if enabled := eval.FindEnabledTransitions(context.Background(), n, &belowThreshold); len(enabled) != 0 {
@@ -787,9 +348,7 @@ func TestEnablementEvaluator_VisitCountGuardEnablesAtThreshold(t *testing.T) {
 		"tok-work": {
 			ID:      "tok-work",
 			PlaceID: "p-init",
-			History: interfaces.TokenHistory{
-				TotalVisits: map[string]int{"review": 3},
-			},
+			History: interfaces.TokenHistory{TotalVisits: map[string]int{"review": 3}},
 		},
 	})
 	enabled := eval.FindEnabledTransitions(context.Background(), n, &atThreshold)
@@ -812,14 +371,10 @@ func TestEnablementEvaluator_LogsNoInputArcs(t *testing.T) {
 	}
 
 	marking := makeTestSnapshot(map[string]*interfaces.Token{})
-
-	enabled := eval.FindEnabledTransitions(context.Background(), n, &marking)
-	if len(enabled) != 0 {
+	if enabled := eval.FindEnabledTransitions(context.Background(), n, &marking); len(enabled) != 0 {
 		t.Fatalf("expected 0 enabled transitions, got %d", len(enabled))
 	}
-
-	entry := logger.findEntry("no input arcs")
-	if entry == nil {
+	if entry := logger.findEntry("no input arcs"); entry == nil {
 		t.Fatal("expected log entry containing 'no input arcs'")
 	}
 }
@@ -844,8 +399,6 @@ func TestEnablementEvaluator_NilLoggerDoesNotPanic(t *testing.T) {
 	marking := makeTestSnapshot(map[string]*interfaces.Token{
 		"tok1": {ID: "tok1", PlaceID: "p1"},
 	})
-
-	// Should not panic with nil logger.
 	enabled := eval.FindEnabledTransitions(context.Background(), n, &marking)
 	if len(enabled) != 1 {
 		t.Fatalf("expected 1 enabled transition, got %d", len(enabled))
@@ -879,313 +432,43 @@ func TestEnablementEvaluator_MultipleTransitions_LogsEach(t *testing.T) {
 		},
 	}
 
-	// Only p1 has a token → t1 enabled, t2 disabled.
 	marking := makeTestSnapshot(map[string]*interfaces.Token{
 		"tok1": {ID: "tok1", PlaceID: "p1"},
 	})
-
 	enabled := eval.FindEnabledTransitions(context.Background(), n, &marking)
 	if len(enabled) != 1 {
 		t.Fatalf("expected 1 enabled transition, got %d", len(enabled))
 	}
-
-	enabledCount := logger.countEntries("transition enabled")
-	disabledCount := logger.countEntries("transition disabled")
-	if enabledCount != 1 {
+	if enabledCount := logger.countEntries("transition enabled"); enabledCount != 1 {
 		t.Errorf("expected 1 'transition enabled' log, got %d", enabledCount)
 	}
-	if disabledCount != 1 {
+	if disabledCount := logger.countEntries("transition disabled"); disabledCount != 1 {
 		t.Errorf("expected 1 'transition disabled' log, got %d", disabledCount)
 	}
 }
 
-func TestEnablementEvaluator_ContextPassedThrough(t *testing.T) {
-	// Verify the evaluator accepts context without error. Context is currently
-	// threaded through for future use (e.g., cancellation, tracing).
-	eval := NewEnablementEvaluator(nil)
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	n := &state.Net{
-		Transitions: map[string]*petri.Transition{},
-	}
-	marking := makeTestSnapshot(map[string]*interfaces.Token{})
-
-	enabled := eval.FindEnabledTransitions(ctx, n, &marking)
-	if len(enabled) != 0 {
-		t.Fatalf("expected 0 enabled transitions, got %d", len(enabled))
-	}
-}
-
-func TestEnablementEvaluator_UsesInjectedClockForCronTimeWindowGuard(t *testing.T) {
-	base := time.Date(2026, 4, 18, 12, 0, 0, 0, time.UTC)
-	dueAt := base.Add(2 * time.Minute)
-	expiresAt := base.Add(7 * time.Minute)
-	currentTime := dueAt.Add(-time.Nanosecond)
-	eval := NewEnablementEvaluator(nil, WithEnablementClock(func() time.Time {
-		return currentTime
-	}))
-
-	n := &state.Net{
+func sameNameGuardNet() *state.Net {
+	return &state.Net{
 		Places: map[string]*petri.Place{
-			interfaces.SystemTimePendingPlaceID: {ID: interfaces.SystemTimePendingPlaceID},
+			"plan:ready": {ID: "plan:ready"},
+			"task:ready": {ID: "task:ready"},
 		},
 		Transitions: map[string]*petri.Transition{
-			"cron-refresh": {
-				ID:         "cron-refresh",
-				WorkerType: "script",
+			"match-items": {
+				ID:   "match-items",
+				Name: "match-items",
 				InputArcs: []petri.Arc{
+					{ID: "plan-in", Name: "plan", PlaceID: "plan:ready", Direction: petri.ArcInput, Cardinality: petri.ArcCardinality{Mode: petri.CardinalityOne}},
 					{
-						ID:          "cron-time",
-						Name:        "time",
-						PlaceID:     interfaces.SystemTimePendingPlaceID,
+						ID:          "task-in",
+						Name:        "task",
+						PlaceID:     "task:ready",
 						Direction:   petri.ArcInput,
 						Cardinality: petri.ArcCardinality{Mode: petri.CardinalityOne},
-						Guard:       &petri.CronTimeWindowGuard{Workstation: "refresh"},
+						Guard:       &petri.SameNameGuard{MatchBinding: "plan"},
 					},
 				},
 			},
 		},
 	}
-	marking := petri.MarkingSnapshot{
-		Tokens: map[string]*interfaces.Token{
-			"time-refresh": schedulerCronTimeToken("time-refresh", "refresh", dueAt, expiresAt),
-		},
-		PlaceTokens: map[string][]string{
-			interfaces.SystemTimePendingPlaceID: {"time-refresh"},
-		},
-	}
-
-	if enabled := eval.FindEnabledTransitions(context.Background(), n, &marking); len(enabled) != 0 {
-		t.Fatalf("enabled before due = %d, want 0", len(enabled))
-	}
-
-	currentTime = dueAt
-	if enabled := eval.FindEnabledTransitions(context.Background(), n, &marking); len(enabled) != 1 {
-		t.Fatalf("enabled at due = %d, want 1", len(enabled))
-	}
-
-	currentTime = expiresAt.Add(-time.Nanosecond)
-	if enabled := eval.FindEnabledTransitions(context.Background(), n, &marking); len(enabled) != 1 {
-		t.Fatalf("enabled before expiry = %d, want 1", len(enabled))
-	}
-
-	currentTime = expiresAt
-	if enabled := eval.FindEnabledTransitions(context.Background(), n, &marking); len(enabled) != 0 {
-		t.Fatalf("enabled at expiry = %d, want 0", len(enabled))
-	}
-}
-
-func TestEnablementEvaluator_OrdersEnabledTransitionsByID(t *testing.T) {
-	eval := NewEnablementEvaluator(nil)
-
-	n := &state.Net{
-		Places: map[string]*petri.Place{
-			"p-alpha": {ID: "p-alpha"},
-			"p-beta":  {ID: "p-beta"},
-			"p-zeta":  {ID: "p-zeta"},
-		},
-		Transitions: map[string]*petri.Transition{
-			"transition-zeta": {
-				ID:         "transition-zeta",
-				WorkerType: "script",
-				InputArcs: []petri.Arc{
-					{ID: "arc-zeta", Name: "work", PlaceID: "p-zeta", Direction: petri.ArcInput, Cardinality: petri.ArcCardinality{Mode: petri.CardinalityOne}},
-				},
-			},
-			"transition-alpha": {
-				ID:         "transition-alpha",
-				WorkerType: "script",
-				InputArcs: []petri.Arc{
-					{ID: "arc-alpha", Name: "work", PlaceID: "p-alpha", Direction: petri.ArcInput, Cardinality: petri.ArcCardinality{Mode: petri.CardinalityOne}},
-				},
-			},
-			"transition-beta": {
-				ID:         "transition-beta",
-				WorkerType: "script",
-				InputArcs: []petri.Arc{
-					{ID: "arc-beta", Name: "work", PlaceID: "p-beta", Direction: petri.ArcInput, Cardinality: petri.ArcCardinality{Mode: petri.CardinalityOne}},
-				},
-			},
-		},
-	}
-	marking := petri.MarkingSnapshot{
-		Tokens: map[string]*interfaces.Token{
-			"tok-zeta":  {ID: "tok-zeta", PlaceID: "p-zeta"},
-			"tok-alpha": {ID: "tok-alpha", PlaceID: "p-alpha"},
-			"tok-beta":  {ID: "tok-beta", PlaceID: "p-beta"},
-		},
-		PlaceTokens: map[string][]string{
-			"p-zeta":  {"tok-zeta"},
-			"p-alpha": {"tok-alpha"},
-			"p-beta":  {"tok-beta"},
-		},
-	}
-
-	for i := 0; i < 10; i++ {
-		enabled := eval.FindEnabledTransitions(context.Background(), n, &marking)
-		got := transitionIDs(enabled)
-		want := []string{"transition-alpha", "transition-beta", "transition-zeta"}
-		if strings.Join(got, ",") != strings.Join(want, ",") {
-			t.Fatalf("iteration %d enabled transition order = %v, want %v", i, got, want)
-		}
-	}
-}
-
-func TestEnablementEvaluator_SelectsOrdinaryTokensByStableID(t *testing.T) {
-	eval := NewEnablementEvaluator(nil)
-	n := &state.Net{
-		Places: map[string]*petri.Place{
-			"p-work": {ID: "p-work"},
-		},
-		Transitions: map[string]*petri.Transition{
-			"transition-work": {
-				ID:         "transition-work",
-				WorkerType: "script",
-				InputArcs: []petri.Arc{
-					{ID: "arc-work", Name: "work", PlaceID: "p-work", Direction: petri.ArcInput, Cardinality: petri.ArcCardinality{Mode: petri.CardinalityN, Count: 2}},
-				},
-			},
-		},
-	}
-	marking := petri.MarkingSnapshot{
-		Tokens: map[string]*interfaces.Token{
-			"tok-c": {ID: "tok-c", PlaceID: "p-work", Color: interfaces.TokenColor{DataType: interfaces.DataTypeWork}},
-			"tok-a": {ID: "tok-a", PlaceID: "p-work", Color: interfaces.TokenColor{DataType: interfaces.DataTypeWork}},
-			"tok-b": {ID: "tok-b", PlaceID: "p-work", Color: interfaces.TokenColor{DataType: interfaces.DataTypeWork}},
-		},
-		PlaceTokens: map[string][]string{
-			"p-work": {"tok-c", "tok-a", "tok-b"},
-		},
-	}
-
-	enabled := eval.FindEnabledTransitions(context.Background(), n, &marking)
-	if len(enabled) != 1 {
-		t.Fatalf("enabled transitions = %d, want 1", len(enabled))
-	}
-	got := tokenIDs(enabled[0].Bindings["work"])
-	want := []string{"tok-a", "tok-b"}
-	if strings.Join(got, ",") != strings.Join(want, ",") {
-		t.Fatalf("bound ordinary tokens = %v, want %v", got, want)
-	}
-}
-
-func TestEnablementEvaluator_SelectsResourceTokensByStableID(t *testing.T) {
-	eval := NewEnablementEvaluator(nil)
-	n := &state.Net{
-		Places: map[string]*petri.Place{
-			"slot:available": {ID: "slot:available"},
-		},
-		Transitions: map[string]*petri.Transition{
-			"transition-slot": {
-				ID:         "transition-slot",
-				WorkerType: "script",
-				InputArcs: []petri.Arc{
-					{ID: "arc-slot", Name: "slot", PlaceID: "slot:available", Direction: petri.ArcInput, Cardinality: petri.ArcCardinality{Mode: petri.CardinalityOne}},
-				},
-			},
-		},
-	}
-	marking := petri.MarkingSnapshot{
-		Tokens: map[string]*interfaces.Token{
-			"slot-2": {ID: "slot-2", PlaceID: "slot:available", Color: interfaces.TokenColor{DataType: interfaces.DataTypeResource}},
-			"slot-1": {ID: "slot-1", PlaceID: "slot:available", Color: interfaces.TokenColor{DataType: interfaces.DataTypeResource}},
-		},
-		PlaceTokens: map[string][]string{
-			"slot:available": {"slot-2", "slot-1"},
-		},
-	}
-
-	enabled := eval.FindEnabledTransitions(context.Background(), n, &marking)
-	if len(enabled) != 1 {
-		t.Fatalf("enabled transitions = %d, want 1", len(enabled))
-	}
-	got := tokenIDs(enabled[0].Bindings["slot"])
-	want := []string{"slot-1"}
-	if strings.Join(got, ",") != strings.Join(want, ",") {
-		t.Fatalf("bound resource tokens = %v, want %v", got, want)
-	}
-}
-
-func TestEnablementEvaluator_ExpandsRepeatedWorkAndResourceBindingsForSameTransition(t *testing.T) {
-	eval := NewEnablementEvaluator(nil)
-	n := &state.Net{
-		Places: map[string]*petri.Place{
-			"task:init":               {ID: "task:init"},
-			"executor-slot:available": {ID: "executor-slot:available"},
-		},
-		Transitions: map[string]*petri.Transition{
-			"process": {
-				ID:         "process",
-				WorkerType: "processor",
-				InputArcs: []petri.Arc{
-					{ID: "work-in", Name: "work", PlaceID: "task:init", Direction: petri.ArcInput, Cardinality: petri.ArcCardinality{Mode: petri.CardinalityOne}, Guard: &petri.DependencyGuard{}},
-					{ID: "slot-in", Name: "slot", PlaceID: "executor-slot:available", Direction: petri.ArcInput, Cardinality: petri.ArcCardinality{Mode: petri.CardinalityN, Count: 1}},
-				},
-			},
-		},
-	}
-	marking := petri.MarkingSnapshot{
-		Tokens: map[string]*interfaces.Token{
-			"work-b": {ID: "work-b", PlaceID: "task:init", Color: interfaces.TokenColor{DataType: interfaces.DataTypeWork}},
-			"work-a": {ID: "work-a", PlaceID: "task:init", Color: interfaces.TokenColor{DataType: interfaces.DataTypeWork}},
-			"slot-2": {ID: "slot-2", PlaceID: "executor-slot:available", Color: interfaces.TokenColor{DataType: interfaces.DataTypeResource}},
-			"slot-1": {ID: "slot-1", PlaceID: "executor-slot:available", Color: interfaces.TokenColor{DataType: interfaces.DataTypeResource}},
-		},
-		PlaceTokens: map[string][]string{
-			"task:init":               {"work-b", "work-a"},
-			"executor-slot:available": {"slot-2", "slot-1"},
-		},
-	}
-
-	enabled := eval.FindEnabledTransitions(context.Background(), n, &marking)
-	if len(enabled) != 1 {
-		t.Fatalf("base enabled candidates = %d, want 1", len(enabled))
-	}
-	expanded := ExpandRepeatedBindings(n, &marking, enabled)
-	if len(expanded) != 2 {
-		t.Fatalf("expanded candidates = %d, want 2", len(expanded))
-	}
-	gotFirst := append(tokenIDs(expanded[0].Bindings["work"]), tokenIDs(expanded[0].Bindings["slot"])...)
-	gotSecond := append(tokenIDs(expanded[1].Bindings["work"]), tokenIDs(expanded[1].Bindings["slot"])...)
-	if strings.Join(gotFirst, ",") != "work-a,slot-1" {
-		t.Fatalf("first candidate tokens = %v, want [work-a slot-1]", gotFirst)
-	}
-	if strings.Join(gotSecond, ",") != "work-b,slot-2" {
-		t.Fatalf("second candidate tokens = %v, want [work-b slot-2]", gotSecond)
-	}
-}
-
-func schedulerCronTimeToken(id string, workstation string, dueAt time.Time, expiresAt time.Time) *interfaces.Token {
-	return &interfaces.Token{
-		ID:      id,
-		PlaceID: interfaces.SystemTimePendingPlaceID,
-		Color: interfaces.TokenColor{
-			WorkID:     id,
-			WorkTypeID: interfaces.SystemTimeWorkTypeID,
-			DataType:   interfaces.DataTypeWork,
-			Tags: map[string]string{
-				interfaces.TimeWorkTagKeySource:          interfaces.TimeWorkSourceCron,
-				interfaces.TimeWorkTagKeyCronWorkstation: workstation,
-				interfaces.TimeWorkTagKeyDueAt:           dueAt.Format(time.RFC3339Nano),
-				interfaces.TimeWorkTagKeyExpiresAt:       expiresAt.Format(time.RFC3339Nano),
-			},
-		},
-	}
-}
-
-func transitionIDs(enabled []interfaces.EnabledTransition) []string {
-	ids := make([]string, len(enabled))
-	for i := range enabled {
-		ids[i] = enabled[i].TransitionID
-	}
-	return ids
-}
-
-func tokenIDs(tokens []interfaces.Token) []string {
-	ids := make([]string, len(tokens))
-	for i := range tokens {
-		ids[i] = tokens[i].ID
-	}
-	return ids
 }
